requirements.txt
# Django & DRF
Django==4.2.10
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0

# Database
psycopg2-binary==2.9.9

# Redis (for caching, optional)
redis==5.0.1

# Environment variables
python-decouple==3.8

# CORS
django-cors-headers==4.3.1

# Image processing
Pillow==10.2.0

# Celery
celery==5.3.4
django-celery-beat==2.5.0

# Testing
pytest==7.4.3
pytest-django==4.7.0
pytest-cov==4.1.0
